+++ typescript/sdk/src/ethers/retry-provider.ts
@@ -5,39 +5,14 @@
 import { BaseProvider, JsonRpcProvider } from '@ethersproject/providers';
 import { ethers } from 'ethers';
 
+import { retryAsync } from '@abacus-network/utils/dist/src/utils';
+
 export type RetryOptions = {
   // The wait interval in between
   interval: number;
 
   // Maximum number of times to rety
   retryLimit: number;
-};
-
-type InFunction<T extends any[], U> = (...params: T) => Promise<U>;
-export function sleep(ms: number): Promise<void> {
-  return new Promise<void>((resolve) => setTimeout(resolve, ms));
-}
-
-// Retries an async function when it raises an exeption
-// if all the tries fail it raises the last thrown exeption
-export const retryAsync = async <InputArgs extends any[], ReturnType>(
-  inFunction: InFunction<InputArgs, ReturnType>,
-  tries: number,
-  params: InputArgs,
-  delay = 100,
-): Promise<ReturnType> => {
-  let saveError;
-  for (let i = 0; i < tries; i++) {
-    try {
-      // it awaits otherwise it'd always do all the retries
-      return await inFunction(...params);
-    } catch (error) {
-      await sleep(delay);
-      saveError = error;
-    }
-  }
-
-  throw saveError;
 };
 
 export class RetryProvider extends BaseProvider {
@@ -53,9 +28,8 @@ export class RetryProvider extends BaseProvider {
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   perform(method: string, params: any): Promise<any> {
     return retryAsync(
-      async () => this.provider.perform(method, params),
+      () => this.provider.perform(method, params),
       this.retryOptions.retryLimit,
-      [],
       this.retryOptions.interval,
     );
   }
@@ -74,7 +48,6 @@ export class RetryJsonRpcProvider extends JsonRpcProvider {
     return retryAsync(
       async () => this.provider.send(method, params),
       this.retryOptions.retryLimit,
-      [],
       this.retryOptions.interval,
     );
   }
